move parserawposts for reddit to the reddit folder. filter out posts without urls. move the logging elsewhere
