+++ src/lemmy/model/Post.ts
@@ -2,5 +2,5 @@ export interface Post {
   id: string;
   title: string;
   content?: string;
-  url?: string;
+  url: string;
 }
