--- src/lemmy/service/parseRawPosts.ts
@@ -1,19 +1,0 @@
-import { RawPost } from '../../reddit/model/RawPost';
-import { Post } from '../model/Post';
-
-export function parseRawPosts(rawPosts: RawPost[]): Post[] {
-  const posts = rawPosts.map((rawPost) => {
-    const post: Post = {
-      id: rawPost.data.id,
-      title: rawPost.data.title,
-      url: rawPost.data.url_overridden_by_dest,
-    };
-
-    if (rawPost.data.selftext.length) {
-      post.content = rawPost.data.selftext;
-    }
-    return post;
-  });
-
-  return posts;
-}
